+++ src/pygaps/__init__.py
@@ -3,7 +3,7 @@
 
 __author__ = 'Paul A. Iacomi'
 __docformat__ = 'restructuredtext'
-__version__ = '1.3.0'
+__version__ = '1.4.0'
 
 # isort:skip_file