+++ tests/classes/test_adsorbate.py
@@ -2,6 +2,8 @@
 This test module has tests relating to the adsorbate class
 """
 
+import warnings
+
 import pytest
 
 import pygaps
@@ -13,7 +15,7 @@ class TestAdsorbate(object):
     Tests the adsorbate class
     """
 
-    def test_adsorbate_created(self, adsorbate_data, basic_adsorbate):
+    def test_adsorbate_create(self, adsorbate_data, basic_adsorbate):
         "Checks adsorbate can be created from test data"
 
         assert adsorbate_data == basic_adsorbate.to_dict()
@@ -21,9 +23,7 @@ class TestAdsorbate(object):
         with pytest.raises(pygaps.ParameterError):
             pygaps.Adsorbate()
 
-        return
-
-    def test_adsorbate_retreived_list(self, adsorbate_data, basic_adsorbate):
+    def test_adsorbate_retrieved_list(self, adsorbate_data, basic_adsorbate):
         "Checks adsorbate can be retrieved from master list"
         pygaps.data.ADSORBATE_LIST.append(basic_adsorbate)
         uploaded_adsorbate = pygaps.Adsorbate.from_list(
@@ -34,22 +34,16 @@ class TestAdsorbate(object):
         with pytest.raises(pygaps.ParameterError):
             pygaps.Adsorbate.from_list('noname')
 
-        return
-
     def test_adsorbate_get_properties(self, adsorbate_data, basic_adsorbate):
         "Checks if properties of a adsorbate can be located"
 
-        assert basic_adsorbate.get_prop(
-            'common_name') == adsorbate_data.get('common_name')
-        assert basic_adsorbate.common_name(
-        ) == adsorbate_data.get('common_name')
+        assert basic_adsorbate.get_prop('common_name') == adsorbate_data.get('common_name')
+        assert basic_adsorbate.common_name() == adsorbate_data.get('common_name')
 
         name = basic_adsorbate.properties.pop('common_name')
         with pytest.raises(pygaps.ParameterError):
             basic_adsorbate.common_name()
         basic_adsorbate.properties['common_name'] = name
-
-        return
 
     @pytest.mark.parametrize('calculated', [True, False])
     def test_adsorbate_named_props(self, adsorbate_data, basic_adsorbate, calculated):
@@ -62,14 +56,33 @@ class TestAdsorbate(object):
             adsorbate_data.get('surface_tension'), 0.001)
         assert basic_adsorbate.liquid_density(temp, calculate=calculated) == pytest.approx(
             adsorbate_data.get('liquid_density'), 0.001)
+        assert basic_adsorbate.gas_density(temp, calculate=calculated) == pytest.approx(
+            adsorbate_data.get('gas_density'), 0.001)
         assert basic_adsorbate.enthalpy_liquefaction(temp, calculate=calculated) == pytest.approx(
             adsorbate_data.get('enthalpy_liquefaction'), 0.001)
 
-        return
+    @pytest.mark.parametrize('calculated, error',
+                             [(True, pygaps.CalculationError),
+                              (False, pygaps.ParameterError)])
+    def test_adsorbate_miss_named_props(self, calculated, error):
+        temp = 77.355
+        ads = pygaps.Adsorbate(nick='n', formula='C2')
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            with pytest.raises(error):
+                ads.molar_mass(calculated)
+            with pytest.raises(error):
+                ads.saturation_pressure(temp, calculate=calculated)
+            with pytest.raises(error):
+                ads.surface_tension(temp, calculate=calculated)
+            with pytest.raises(error):
+                ads.liquid_density(temp, calculate=calculated)
+            with pytest.raises(error):
+                ads.gas_density(temp, calculate=calculated)
+            with pytest.raises(error):
+                ads.enthalpy_liquefaction(temp, calculate=calculated)
 
     def test_adsorbate_print(self, basic_adsorbate):
         """Checks the printing is done"""
 
         print(basic_adsorbate)
-
-        return