+++ CHANGELOG.rst
@@ -3,6 +3,20 @@ Changelog
 =========
 
 
+1.4.0 (2018-11-10)
+------------------
+
+Features:
+
+* Added the GAB isotherm model
+
+Bugfixes:
+
+* Improved pore size distribution calculations to display cumulative pore
+  volume when called.
+* Fixed the "all-nol" selection parameter for legend display in isotherm
+  graphs.
+
 1.3.0 (2018-08-13)
 ------------------